requirements.txt
# Django and REST Framework
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0

# Machine Learning and Image Processing
onnxruntime==1.21.1
numpy<2
Pillow==10.2.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
opencv-python==4.8.0.76
ultralytics==8.0.196
dill==0.3.6
soundfile==0.12.1
pydub==0.25.1
librosa

# TensorFlow for compatibility with the model
tensorflow==2.18.0

# Transformers & RAG
transformers
huggingface-hub
sentence-transformers
faiss-cpu
accelerate
pypdf
langchain==0.3.25
langchain-community==0.3.23
langchain-groq>=0.3.2
langchain-huggingface==0.1.0
elevenlabs
tenacity

# Development Tools
python-dotenv==1.0.0
ipywidgets
jupyterlab

# Additional Libraries
bitsandbytes>=0.39.1
streamlit>=1.25.0
PyPDF2
datasets>=2.14.5
tokenizers
lime
channels
daphne
